Track inventory purchases, invoice stock movements and returns

+++ database/migrations/2024_02_03_064819_create_expenses_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('title')->nullable();
             $table->date('date')->nullable();
-            $table->integer('amount')->default(0);
+            $table->decimal('amount', 15, 2)->default(0);
             $table->string('receipt')->nullable();
             $table->text('notes')->nullable();
             $table->integer('parent_id')->default(0);
